Keep task buttons in the sidebar footer
