+++ src/components/sidebar/Sidebar.vue
@@ -8,23 +8,65 @@
         &#x25B2;
       </span>
     </div>
+
     <div class="lv-sidebar-content" v-show="!contentCollapsed">
+      <h4>{{ task.objective }}</h4>
+      <p>{{ task.guidance }}</p>
       <slot></slot>
     </div>
+
+    <div class="lv-sidebar-footer">
+      <button
+        v-if="showNote"
+        class="btn btn-block"
+        @click="toggleCollapseNote">
+        Add a note
+      </button>
+
+      <transition name="fade-height"
+        v-show="showNote"
+        v-if="!collapseNote">
+        <textarea
+          ref="note"
+          rows="3"
+          placeholder="Leave a note..."
+          v-model="note"
+          @input="updateNote">
+        </textarea>
+      </transition>
+
+      <button
+        class="btn btn-block btn-green"
+        @click="submit">
+        Submit
+      </button>
+    </div>
+
   </div>
 </template>
 
 <script>
+import Task from '@/task'
+
 export default {
   data: function () {
     return {
-      contentCollapsed: false
+      contentCollapsed: false,
+      collapseNote: true
     }
   },
 
   props: {
     title: {
       type: String,
+      default: 'Task'
+    },
+    task: {
+      type: Task,
+      required: true
+    },
+    showNote: {
+      type: Boolean,
       required: true
     }
   },
@@ -35,18 +77,46 @@ export default {
         toggle: true,
         active: this.contentCollapsed
       }
+    },
+    note: function () {
+      const comments = this.task.getAnnotationsByMotivation('commenting')
+      return comments.length ? comments[0].body.value : ''
     }
   },
 
   methods: {
     toggleCollapse () {
       this.contentCollapsed = !this.contentCollapsed
+    },
+    /**
+     * Toggle the collapsing of the note input.
+     */
+    toggleCollapseNote () {
+      this.collapseNote = !this.collapseNote
+    },
+
+    /**
+     * Emit the noteupdated event with the note value.
+     */
+    updateNote (evt) {
+      this.$emit('noteupdated', this.task, evt.target.value)
+    },
+
+    /**
+     * Emit the submit event with the task.
+     */
+    submit () {
+      this.$emit('submit', this.task)
     }
   }
 }
 </script>
 
 <style lang="scss" scoped>
+@import '../../assets/style/settings';
+@import '../../assets/style/partials/buttons';
+@import '../../assets/style/partials/transitions';
+
 .lv-sidebar {
   font-size: 14px;
   color: #FFFFFF;
@@ -55,6 +125,8 @@ export default {
   border: 2px solid rgb(85, 85, 85);
   background-color: rgba(0, 0, 0, 0.75);
   margin: 1rem;
+  display: flex;
+  flex-direction: column;
 
   @media screen and (min-width: 992px) {
     flex-direction: column;
@@ -80,6 +152,23 @@ export default {
     position: relative;
     padding: 0.6em;
     overflow-y: auto;
+
+    h4 {
+      margin: 0;
+    }
+  }
+
+  .lv-sidebar-footer {
+    padding: 0.6rem;
+    border-top: 1px solid $gray;
+
+    button:not(:last-child) {
+      margin-bottom: 0.6rem;
+    }
+
+    textarea {
+      margin-top: 0.6rem;
+    }
   }
 
   .toggle {
